Revert "CXBOX-89 | Primary in AssocListPopup"
This reverts commit 49c45e45675460d3ceb3ef2a25838b12f9a120bc.

Revert "CXBOX-89 | fixed meta" frontend part only

This reverts commit 564dfe2df4200262510c33dac83ba9289a69e3aa.

Revert "CXBOX-89 | added title to options.primary"

This reverts commit e5784925148520b5200d4fabb148a15fb97c8c18.
