--- ui/src/components/widgets/AssocListPopup/PassiveAssocListPopup/Title.less
@@ -1,37 +1,0 @@
-.title {
-    font-size: 24px;
-}
-
-.tagArea {
-    overflow-y: auto;
-    max-height: 200px;
-    :global {
-        .ant-tag {
-            position: relative;
-            margin: 2px 2px 0 2px;
-            padding-right: 20px;
-            line-height: 18px;
-            background: #ffffff;
-            white-space: initial;
-
-            .anticon {
-                position: absolute;
-                display: block;
-                top: 3px;
-                right: 6px;
-            }
-        }
-    }
-}
-
-.primary:global(.ant-tag) {
-    color: #fff;
-    background-color: #1890ff;
-    border-color: #1890ff;
-
-    :global(.anticon) {
-        color: #fff;
-    }
-
-}
-
