+++ ui/src/components/widgets/DashboardList/DashboardList.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
+import { WidgetTableMeta } from '@cxbox-ui/core/interfaces/widget'
 import Table from '../Table/Table'
-import { AppWidgetTableMeta } from '../../../interfaces/widget'
 
 interface DashboardListProps {
-    meta: AppWidgetTableMeta
+    meta: WidgetTableMeta
 }
 
 function DashboardList({ meta }: DashboardListProps) {
