+++ ui/src/components/widgets/PickListPopup/PickListPopup.tsx
@@ -1,16 +1,16 @@
 import React from 'react'
 import cn from 'classnames'
 import { PickListPopup as CorePickListPopup } from '@cxbox-ui/core'
+import { WidgetTableMeta } from '@cxbox-ui/core/interfaces/widget'
 import tableStyles from '../Table/Table.less'
 import styles from './PickListPopup.module.css'
 import Pagination from '../../ui/Pagination/Pagination'
 import { AppState } from '../../../interfaces/storeSlices'
 import { useSelector } from 'react-redux'
 import AssocListPopup from '../AssocListPopup/AssocListPopup'
-import { AppWidgetTableMeta } from '../../../interfaces/widget'
 
 interface PickListPopupProps {
-    meta: AppWidgetTableMeta
+    meta: WidgetTableMeta
 }
 
 function PickListPopup({ meta }: PickListPopupProps) {
