--- ui/src/hooks/useAppSelector.ts
@@ -1,4 +1,0 @@
-import { TypedUseSelectorHook, useSelector } from 'react-redux'
-import { AppState } from '../interfaces/storeSlices'
-
-export const useAppSelector: TypedUseSelectorHook<AppState> = useSelector
